split updated

+++ ft_strmapi.c
@@ -6,17 +6,17 @@
 /*   By: mmunoz-f <mmunoz-f@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/23 17:08:21 by mmunoz-f          #+#    #+#             */
-/*   Updated: 2022/01/10 18:15:12 by mmunoz-f         ###   ########.fr       */
+/*   Updated: 2022/01/11 16:25:44 by mmunoz-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strmapi(const char *src, char (*f)(size_t, char))
+char	*ft_strmapi(const char *src, char (*f)(unsigned int, char))
 {
-	char	*a;
-	size_t	i;
-	size_t	len;
+	char			*a;
+	unsigned int	i;
+	size_t			len;
 
 	if (!src)
 		return (0);
